+++ src/GLOBAL_TEXT_BG/text.ts
@@ -1,0 +1,3 @@
+export const textColor = 'white';
+export const subTextColorBlack = '#B0B0B0';
+export const subTextColorWhite = 'white';
