+++ src/main.jsx
@@ -1,10 +1,13 @@
-import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
 import App from './App.jsx'
-
+import { Provider } from 'react-redux';
+import { StrictMode } from 'react';
+import { store } from './redux/store.js';
 createRoot(document.getElementById('root')).render(
   <StrictMode>
+  <Provider store={store}>
     <App />
-  </StrictMode>,
+  </Provider>
+</StrictMode>
 )
