+++ src/redux/chatting/Chatting.ts
@@ -1,0 +1,30 @@
+// src/store/userSlice.ts
+import { createSlice, PayloadAction } from '@reduxjs/toolkit';
+
+interface UserState {
+  name: string;
+  loggedIn: boolean;
+}
+
+const initialState: UserState = {
+  name: '',
+  loggedIn: false,
+};
+
+const userSlice = createSlice({
+  name: 'user',
+  initialState,
+  reducers: {
+    login(state, action: PayloadAction<string>) {
+      state.name = action.payload;
+      state.loggedIn = true;
+    },
+    logout(state) {
+      state.name = '';
+      state.loggedIn = false;
+    },
+  },
+});
+
+export const { login, logout } = userSlice.actions;
+export default userSlice.reducer;
